refactor(console): extract the onboarding app from the admin console
