--- packages/console/src/cloud/pages/Onboarding/index.module.scss
@@ -1,6 +1,0 @@
-.onBoarding {
-  height: 100%;
-  overflow: hidden;
-  display: flex;
-  flex-direction: column;
-}
