--- packages/console/src/cloud/scss/cloud-page-size.scss
@@ -1,1 +1,0 @@
-$questionnaire-content-width: 858px;
