--- packages/console/src/cloud/utils.ts
@@ -1,4 +1,0 @@
-import type { OnboardingPage } from './types';
-import { CloudRoute } from './types';
-
-export const getOnboardPagePathname = (page: OnboardingPage) => `/${CloudRoute.Onboarding}/${page}`;
