+++ packages/console/src/containers/AppContent/components/Topbar/index.tsx
@@ -2,19 +2,20 @@ import classNames from 'classnames';
 import { useTranslation } from 'react-i18next';
 
 import Logo from '@/assets/images/logo.svg';
-import EarlyBirdGift from '@/cloud/components/EarlyBirdGift';
 import Spacer from '@/components/Spacer';
 import { isCloud } from '@/consts/cloud';
+import EarlyBirdGift from '@/onboarding/components/EarlyBirdGift';
 import GetStartedProgress from '@/pages/GetStarted/components/GetStartedProgress';
 
 import UserInfo from '../UserInfo';
 import * as styles from './index.module.scss';
 
 type Props = {
+  isLogoOnly?: boolean;
   className?: string;
 };
 
-const Topbar = ({ className }: Props) => {
+const Topbar = ({ isLogoOnly = false, className }: Props) => {
   const { t } = useTranslation(undefined, { keyPrefix: 'admin_console' });
 
   return (
@@ -23,9 +24,13 @@ const Topbar = ({ className }: Props) => {
       <div className={styles.line} />
       <div className={styles.text}>{t('title')}</div>
       <Spacer />
-      <GetStartedProgress />
-      {isCloud && <EarlyBirdGift />}
-      <UserInfo />
+      {!isLogoOnly && (
+        <>
+          <GetStartedProgress />
+          {isCloud && <EarlyBirdGift />}
+          <UserInfo />
+        </>
+      )}
     </div>
   );
 };
