+++ packages/console/src/onboarding/App.module.scss
@@ -1,0 +1,6 @@
+@use '@/scss/underscore' as _;
+
+.app {
+  position: absolute;
+  inset: 0;
+}
