+++ packages/console/src/onboarding/components/ActionBar/index.module.scss
