+++ packages/console/src/onboarding/components/ActionBar/index.tsx
