+++ packages/console/src/onboarding/components/CardSelector/MultiCardSelector/index.module.scss
