+++ packages/console/src/onboarding/components/CardSelector/types.ts
