+++ packages/console/src/onboarding/components/EarlyBirdGift/GiftModal/index.module.scss
