+++ packages/console/src/onboarding/components/EarlyBirdGift/GiftModal/index.tsx
