+++ packages/console/src/onboarding/components/ProgressBar/index.module.scss
