+++ packages/console/src/onboarding/components/ProgressBar/index.tsx
