+++ packages/console/src/onboarding/constants/index.ts
