+++ packages/console/src/onboarding/containers/AppContent/index.module.scss
@@ -1,0 +1,13 @@
+.app {
+  position: absolute;
+  inset: 0;
+  display: flex;
+  flex-direction: column;
+}
+
+.content {
+  height: 100%;
+  overflow: hidden;
+  display: flex;
+  flex-direction: column;
+}
