+++ packages/console/src/onboarding/pages/About/index.module.scss
