+++ packages/console/src/onboarding/pages/Congrats/index.module.scss
