+++ packages/console/src/onboarding/pages/SignInExperience/components/PlatformTabs/PlatformTab.module.scss
