+++ packages/console/src/onboarding/pages/SignInExperience/components/PlatformTabs/index.tsx
