+++ packages/console/src/onboarding/pages/SignInExperience/components/Preview/index.module.scss
