+++ packages/console/src/onboarding/pages/Welcome/index.module.scss
