+++ packages/console/src/onboarding/pages/Welcome/options.tsx
@@ -2,7 +2,7 @@ import Building from '@/assets/images/building.svg';
 import Cloud from '@/assets/images/cloud.svg';
 import Database from '@/assets/images/database.svg';
 import Pizza from '@/assets/images/pizza.svg';
-import type { Option as SelectorOption } from '@/cloud/components/CardSelector';
+import type { Option as SelectorOption } from '@/onboarding/components/CardSelector';
 
 import { DeploymentType, Project } from '../../types';
